Move into the normal directory tree
