+++ composer.json
@@ -31,8 +31,7 @@
         "square/connect": "2.20190710.0",
         "subfission/cas": "dev-master",
         "uxweb/sweet-alert": "2.0.1",
-        "vyuldashev/nova-permission": "1.9.0",
-        "robojackets/apiary-nova-collect-attendance": "*"
+        "vyuldashev/nova-permission": "1.9.0"
     },
     "require-dev": {
         "beyondcode/laravel-dump-server": "1.3.0",
@@ -82,10 +81,6 @@
         {
             "type": "composer",
             "url": "https://nova.laravel.com"
-        },
-        {
-            "type": "path",
-            "url": "./nova-components/CollectAttendance"
         }
     ]
 }
