--- nova-components/CollectAttendance/.gitignore
@@ -1,10 +1,0 @@
-/.idea
-/vendor
-/node_modules
-package-lock.json
-composer.phar
-composer.lock
-phpunit.xml
-.phpunit.result.cache
-.DS_Store
-Thumbs.db
