--- nova-components/CollectAttendance/composer.json
@@ -1,29 +1,0 @@
-{
-    "name": "robojackets/apiary-nova-collect-attendance",
-    "description": "A Laravel Nova resource tool.",
-    "keywords": [
-        "laravel",
-        "nova"
-    ],
-    "license": "MIT",
-    "require": {
-        "php": ">=7.1.0"
-    },
-    "autoload": {
-        "psr-4": {
-            "Robojackets\\Apiary\\Nova\\CollectAttendance\\": "src/"
-        }
-    },
-    "extra": {
-        "laravel": {
-            "providers": [
-                "Robojackets\\Apiary\\Nova\\CollectAttendance\\ToolServiceProvider"
-            ]
-        }
-    },
-    "config": {
-        "sort-packages": true
-    },
-    "minimum-stability": "dev",
-    "prefer-stable": true
-}
