--- nova-components/CollectAttendance/resources/js/tool.js
@@ -1,5 +1,0 @@
-import Vuelidate from 'vuelidate'
-Nova.booting((Vue, router, store) => {
-    Vue.use(Vuelidate)
-    Vue.component('collect-attendance', require('./components/Tool'))
-})
