--- nova-components/CollectAttendance/resources/sass/tool.scss
@@ -1,1 +1,0 @@
-// Nova Tool CSS
