--- nova-components/CollectAttendance/webpack.mix.js
@@ -1,5 +1,0 @@
-let mix = require('laravel-mix')
-
-mix.setPublicPath('dist')
-    .js('resources/js/tool.js', 'js')
-    .sass('resources/sass/tool.scss', 'css')
